fix: keep control inputs as uint8 in batch mode to preserve normalization
The model's _normalize_video_databatch_inplace checks dtype == uint8
before normalizing from [0,255] to [-1,1]. Converting to bfloat16 early
bypassed this check, causing control inputs to have wrong value range
and poor quality output for all batch samples.

+++ cosmos_transfer2/_src/transfer2/inference/inference_pipeline.py
@@ -806,7 +806,9 @@ class ControlVideo2WorldInference:
             if all(c is not None for c in control_list):
                 # Stack along new dim=0 to get (B, C, T, H, W)
                 control_batch = torch.stack(control_list, dim=0)
-                data_batch[control_key] = control_batch.to(dtype=torch.bfloat16, device="cuda")
+                # Keep as uint8! Model's _normalize_video_databatch_inplace expects uint8 and
+                # will normalize to [-1, 1]. Converting to bfloat16 here bypasses normalization.
+                data_batch[control_key] = control_batch.to(device="cuda")
                 loaded_control_keys.add(key)
                 # Add mask for this control
                 mask_key = f"{control_key}_mask"
